Make `proxy_protocol_networks` variable read-write.
